lavague_application/requirements.txt
black
beautifulsoup4
lavague
lavague-core
nltk
python-dotenv
requests
transformers
wikipedia

